main changes

+++ code/DBFile.h
@@ -19,7 +19,7 @@ private:
 	//currPage turned to Page for simplicity
 	Page page;
 	//filecount turned to currpage for simplicity
-	off_t currPage;
+	int currPage;
 
 public:
 	DBFile ();
